added option to interaction importers for skipping pubchem rebuild

+++ lib/perl/Genome/Model/Tools/Dgidb/Import/Pharmgkb.pm
@@ -31,6 +31,13 @@ class Genome::Model::Tools::Dgidb::Import::Pharmgkb {
             default => '/gscmnt/sata132/techd/mgriffit/DruggableGenes/TSV/PharmGKB_WashU_INTERACTIONS.tsv',
             doc => 'PATH.  Path to .tsv file for drug gene interactions',
         },
+        skip_pubchem => {
+            is => 'Boolean',
+            is_input => 1,
+            is_optional => 1,
+            default => 0,
+            doc => 'Skip _destroy_and_rebuild_pubchem_and_drug_groups step',
+        },        
         version => {
             doc => 'VERSION.  Version (date) of release of data files from PharmGKB',
         },
@@ -106,7 +113,9 @@ sub execute {
     my $self = shift;
     $self->input_to_tsv();
     $self->import_tsv();
-    $self->_destroy_and_rebuild_pubchem_and_drug_groups();
+    unless ($self->skip_pubchem){
+        $self->_destroy_and_rebuild_pubchem_and_drug_groups();
+    }
     return 1;
 }
