+++ lib/perl/Genome/Model/Tools/Dgidb/Import/Talc.pm
@@ -21,6 +21,13 @@ class Genome::Model::Tools::Dgidb::Import::Talc {
             default => 0,
             doc => 'Print more output while running',
         },
+        skip_pubchem => {
+            is => 'Boolean',
+            is_input => 1,
+            is_optional => 1,
+            default => 0,
+            doc => 'Skip _destroy_and_rebuild_pubchem_and_drug_groups step',
+        },
         interactions_outfile => {
             is => 'Path',
             is_input => 1,
@@ -105,7 +112,9 @@ sub execute {
     my $self = shift;
     $self->input_to_tsv();
     $self->import_tsv();
-    $self->_destroy_and_rebuild_pubchem_and_drug_groups();
+    unless ($self->skip_pubchem){
+        $self->_destroy_and_rebuild_pubchem_and_drug_groups();
+    }
     return 1;
 }
